chore: updated return type from double to BigInt

+++ packages/ethereum_api/lib/src/wallet/magic_api_client/unsupported.dart
@@ -118,7 +118,7 @@ class MagicWalletApiClient implements MagicApiClient {
   }
 
   @override
-  Future<double> getGasPrice() {
+  Future<BigInt> getGasPrice() {
     // TODO: implement getGasPrice
     throw UnimplementedError(
       'getGasPrice is not implemented on the current platform',
